WinGui: Fix an issue with long chapters causing scans to fail to complete. Fixes #713

+++ win/CS/HandBrake.ApplicationServices/Interop/Json/Scan/Duration.cs
@@ -32,6 +32,6 @@ namespace HandBrake.ApplicationServices.Interop.Json.Scan
         /// <summary>
         /// Gets or sets the ticks.
         /// </summary>
-        public int Ticks { get; set; }
+        public long Ticks { get; set; }
     }
 }
